requirements.txt
tensorflow-gpu==1.8.0
keras==2.2.4
numpy==1.14.2
cv2==4.1.0
scipy==1.2.1
scikit-learn==0.21.2
pywt==1.0.3
 
